3. FCM/requirements.txt
# python -m pip install --upgrade pip
# pip install -r requirements.txt  
firebase_admin
mysql-connector-python
python-dotenv